x86-64: Simplify constant argument code generation
The specialized instruction selection rules are not needed for correctness
because EXPR_VALUE and EXPR_FVALUE combined with EXPR_ARG deals with immediate
arguments just fine.

The immediate floating point rule is broken so drop bot specialized runes for
now. We can reintroduce them later on if we can show there's a performance
benefit in having them.

Signed-off-by: Pekka Enberg <penberg@kernel.org>