x86, jit: Simplify multianewarray code generation
This patch simplifies multianewarray code generation in preparation for fixing
multianewarray on x86-64 and other architectures that pass arguments in
registers.

The main issue with the current code is that arguments passed to
multiarray_size_check() and vm_object_alloc_multi_array() are shared which is
fine on 32-bit x86 where arguments are on the stack. Unfortunately for x86-64
arguments are passed in registers and are thus clobbered by the
multiarray_size_check() function.

By dropping EXPR_MULTIARRAY_SIZE_CHECK, we avoid the issue completely which
makes fixing multianewarray easier.

Signed-off-by: Pekka Enberg <penberg@kernel.org>